pyproject.toml
[project]
name = "pidash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "johang13", email = "chris.grobler@outlook.co.nz" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "openmeteo-requests>=1.7.5",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "requests-cache>=1.3.1",
    "retry-requests>=2.0.0",
]

[project.scripts]
pidash = "pidash:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pylint>=4.0.5",
]
